Split Table elements in dedicated files
