+++ src/sections/BlockInfoSection.tsx
@@ -31,17 +31,12 @@ import LoadingSpinner from '../components/LoadingSpinner'
 import PageSwitch from '../components/PageSwitch'
 import Section from '../components/Section'
 import SectionTitle, { SecondaryTitle } from '../components/SectionTitle'
-import {
-  AnimatedCell,
-  DetailsRow,
-  DetailToggle,
-  HighlightedCell,
-  Row,
-  Table,
-  TableBody,
-  TableHeader,
-  TDStyle
-} from '../components/Table'
+import HighlightedCell from '../components/Table/HighlightedCell'
+import Table, { TDStyle } from '../components/Table/Table'
+import TableBody from '../components/Table/TableBody'
+import { AnimatedCell, DetailToggle, TableDetailsRow } from '../components/Table/TableDetailsRow'
+import TableHeader from '../components/Table/TableHeader'
+import TableRow from '../components/Table/TableRow'
 import Timestamp from '../components/Timestamp'
 import usePageNumber from '../hooks/usePageNumber'
 import useTableDetailsState from '../hooks/useTableDetailsState'
@@ -207,7 +202,7 @@ const TransactionRow: FC<TransactionRowProps> = ({ transaction }) => {
 
   return (
     <>
-      <Row key={t.hash} isActive={detailOpen} onClick={toggleDetail}>
+      <TableRow key={t.hash} isActive={detailOpen} onClick={toggleDetail}>
         <td>
           <TransactionIcon />
         </td>
@@ -231,8 +226,8 @@ const TransactionRow: FC<TransactionRowProps> = ({ transaction }) => {
           />
         </td>
         <DetailToggle isOpen={detailOpen} onClick={toggleDetail} />
-      </Row>
-      <DetailsRow openCondition={detailOpen}>
+      </TableRow>
+      <TableDetailsRow openCondition={detailOpen}>
         <td />
         <td />
         <AnimatedCell>
@@ -249,7 +244,7 @@ const TransactionRow: FC<TransactionRowProps> = ({ transaction }) => {
           {t.outputs && t.outputs.map((o, i) => <Amount value={BigInt(o.amount)} key={i} />)}
         </AnimatedCell>
         <td />
-      </DetailsRow>
+      </TableDetailsRow>
     </>
   )
 }
