+++ src/sections/BlockSection.tsx
@@ -25,11 +25,12 @@ import styled, { css } from 'styled-components'
 
 import { GlobalContext } from '..'
 import { TightLink } from '../components/Links'
-import LoadingSpinner from '../components/LoadingSpinner'
 import PageSwitch from '../components/PageSwitch'
 import Section from '../components/Section'
 import SectionTitle from '../components/SectionTitle'
-import { Table, TableBody, TableHeader, TDStyle } from '../components/Table'
+import Table, { TDStyle } from '../components/Table/Table'
+import TableBody from '../components/Table/TableBody'
+import TableHeader from '../components/Table/TableHeader'
 import Timestamp from '../components/Timestamp'
 import usePageNumber from '../hooks/usePageNumber'
 import { BlockList } from '../types/api'
