+++ src/sections/TransactionInfoSection.tsx
@@ -29,7 +29,9 @@ import { AddressLink, TightLink } from '../components/Links'
 import LoadingSpinner from '../components/LoadingSpinner'
 import Section from '../components/Section'
 import SectionTitle from '../components/SectionTitle'
-import { HighlightedCell, Table, TableBody } from '../components/Table'
+import HighlightedCell from '../components/Table/HighlightedCell'
+import Table from '../components/Table/Table'
+import TableBody from '../components/Table/TableBody'
 import Timestamp from '../components/Timestamp'
 import { APIResp } from '../utils/client'
 import { useInterval } from '../utils/hooks'
